+++ custom_components/ducobox-connectivity-board/model/utils.py
